Change size_t to int in virtual ReadV.

+++ src/IO.cc
@@ -79,8 +79,9 @@ int IO::Read (char *buff, long long off, int size)
  * Perform a readv from the cache
  */
 #if defined(HAVE_READV)
-ssize_t IO::ReadV (const XrdOucIOVec *readV, size_t n)
-{ 
+
+int IO::ReadV (const XrdOucIOVec *readV, int n)
+{
     printf("======== IO::ReadV \n");
     ssize_t bytes_read = 0;
     size_t missing = 0;
